requirements.txt
fastapi
pandas
pydantic
uvicorn
psycopg2
sqlalchemy
python-multipart
python-jose
passlib
bcrypt
python-dotenv